Renaming project to MaRe
