+++ pom.xml
@@ -3,13 +3,13 @@
 
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>se.uu.it</groupId>
-	<artifactId>easymr</artifactId>
-	<version>0.4.0-SNAPSHOT</version>
+	<artifactId>mare</artifactId>
+	<version>0.3.0-SNAPSHOT</version>
 	<name>${project.artifactId}</name>
-	<url>https://github.com/mcapuccini/EasyMapReduce</url>
+	<url>https://github.com/mcapuccini/MaRe</url>
 
 	<description>
-	EasyMapReduce leverages the power of Docker and Spark to run and scale your serial tools in MapReduce fashion.
+	MaRe leverages the power of Docker and Spark to run and scale your serial tools in MapReduce fashion.
 	</description>
 
 	<developers>
@@ -51,9 +51,9 @@
 	</distributionManagement>
 
 	<scm>
-		<connection>scm:git:git://github.com/mcapuccini/EasyMapReduce.git</connection>
-		<developerConnection>scm:git:git@github.com:mcapuccini/EasyMapReduce.git</developerConnection>
-		<url>https://github.com/mcapuccini/EasyMapReduce/tree/master</url>
+		<connection>scm:git:git://github.com/mcapuccini/MaRe.git</connection>
+		<developerConnection>scm:git:git@github.com:mcapuccini/MaRe.git</developerConnection>
+		<url>https://github.com/mcapuccini/MaRe/tree/master</url>
 	</scm>
 
 	<dependencies>
