+++ src/main/scala/se/uu/it/mare/EasyFiles.scala
@@ -1,4 +1,4 @@
-package se.uu.it.easymr
+package se.uu.it.mare
 
 import java.io.File
 import java.io.FileNotFoundException
@@ -8,9 +8,10 @@ import java.util.regex.Pattern
 
 import scala.io.Source
 import scala.util.Properties
+
 import org.apache.log4j.Logger
 
-private[easymr] object EasyFiles {
+private[mare] object EasyFiles {
 
   // Set temporary directory
   private val tmpDir = new File(Properties.envOrElse("TMPDIR", "/tmp"))
@@ -26,7 +27,7 @@ private[easymr] object EasyFiles {
   // Logger
   private lazy val log = Logger.getLogger(getClass.getName)
 
-  private def newTmpFile = new File(tmpDir, "easymr_" + UUID.randomUUID.toString)
+  private def newTmpFile = new File(tmpDir, "mare_" + UUID.randomUUID.toString)
 
   def createTmpFile = {
     val file = EasyFiles.newTmpFile
